backend/main-cloud-run/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.8
gunicorn==20.1.0
google-cloud-speech==2.4.0
google-api-core==1.28.0